+++ Grid/Column/MassActionColumn.php
@@ -15,7 +15,7 @@ class MassActionColumn extends Column
 {
     const ID = '__action';
 
-    private $gridHash;
+    protected $gridHash;
 
     public function __construct($gridHash)
     {
@@ -32,7 +32,7 @@ class MassActionColumn extends Column
     {
         return '<input type="checkbox" class="action" value="1" name="'.$this->gridHash.'['.self::ID.']['.$row->getPrimaryFieldValue().']"/>';
     }
-	    
+
     public function getType()
     {
         return 'massaction';