+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+`8 May 2012`
+
+ * Nothing private!
+
 `7 May 2012`
 
  * Fix bug - Templates are not reset for each grid displayed